Modify: 기존 Message 컴포넌트 ZzalMessage로 네이밍 변경 및 UI 수정
